requirements.txt
# M7 Anti-Gravity Bot - Required Python Packages

# Core Data & Analysis
yfinance>=0.2.28
pandas>=2.0.0
numpy>=1.24.0

# Web Dashboard
streamlit>=1.28.0
plotly>=5.17.0

# API & Networking
requests>=2.31.0

# Environment Variables
python-dotenv>=1.0.0

# Scientific Computing (for potential future strategies)
scipy>=1.11.0
